Base of modular page creation
Setup a settings parent class for other pages to be created from.
Created 3 different pages for the menu with different stuff in them to test.
Changed a bit how page_init works. Before it initiated the sections and registered them, but it was making so every section from every page got initiated on every page. now it just registers the settings and I have another method doing the field creating part.